Finally found a good solution for the accessibility issues!
We're launching a separate instance of the helper independent from launchd and passing it command line args, so it can quickly install itself in system settings and then exit immediately. This works great!

We can still improve the implementation in the next commits but we finally have a basic approach that works!