Turned off code stripping.
And swift stripping.

This was part of an effort to try and debug on older macOS versions. Not sure this is good or useful. I think you need to use .dsym file to get proper debug info, and not stripping the binary doesn't really help.

Not sure what Swift symbol stripping does at all.