Implemented backend to send scrolls with normal non-Hybrid animation curve.
See a few commits ago where we implemented the first part of this change in ScrollConfg

Edit: Stuff we just implemented works fine but stopping momentumScroll at the end of the animation inside Scroll.m doesn't work.  This might be why the non-Inertial scrolling felt weird in the previous MMF 3 Betas!