Lots of experiments to try and get desktop tinting to work properly. Made a huge mess. Will revert this.
The goal was to have it so that the the addField's dropShadow which appears on mouse hover doesn't change the addField's color. But to *also* have desktop tinting.

We found two ways to do this:

1. Use one of the system colors which are naturally opaque but also sopport desktop tinting.

 I've found these four:
- .controlBackgroundColor
- .textBackgroundColor
- .underPageBackgroundColor
- .windowBackgroundColor

2. Do custom drawing for the NSShadow

Problem is that the only way to put a hard cut inside of the the shadow seems to be to render it to an Image alongside the shape that is casting the shadow and then cutting that shape out from the image and then rendering the image. But when drawing the image it can't spill outside of the view's frame like the shadow does. So you have to create another view (or layer) to render the shadow image with the cutout.

Not sure how you would coordinate the drawing between the hostView and shadowDrawing view. Especially since we're animating the size of the hostView with a  transform.

In the end I gave up on this approach because it's just going to take too much time and I found some systemColors that support desktop tinting which also look nice. Awesome! (God I spent too much time on this)

Aside from that this commit also changes some other colors on the customViews on the buttonTab to support desktop tinting better. Also we made the height of the actionTable 1 pixel shorter to prevent visual issues with the new colors (When the actiontable grid overlaps the actiontable border the bordercolor becomes very light and it's very noticalble and looks weird)