Big renaming effort. Anything called `transformation` is now called `remap` instead. Except TransformationUtility which is now called ModificationUtility.
I never really had a clear idea of what a `transformation` is.

Now I have a clear idea for the terminology:

The remaps are a map modifiers → (trigger → effect). Where the (trigger → effect) map is also called a modification.

Any module in the code primarily dealing with the remaps is now called `Remap...` instead of `Transformation...`. We are not using `Transformation` any more at all since it's very vague.

This commit also includes some other minor changes and cleanup.