Turned off swift autobridging for the classes that deal with the remaps dict.
All these classes now expose the native Foundation types instead of very slowly converting between Swift types and ObjC types all the time. This is very cumbersome to achieve but it's so much faster!

The CPU usage from spamming button 4 on the default settings under debug is now around 7.0!% (down from around 8.5%). Instruments says that 75% ( = 5% CPU usage)of that is just CocoaLumberjack so my core code is just 1.75% CPU usage - and that's with a debug build under swift. The performance of the core code was ca. doubled in this scenario. So that's cool.