Various small improvements to UI text. Mostly punctuation. Added lots of commas, primarily to German. Now the lack of closing periods feels slightly off.
Not totally sure which system to follow for the periods.

Material Design recommends only using periods to separate consecutive sentences (Src: https://ux.stackexchange.com/questions/122147/periods-in-ux-ui-microcopy). And never elsewhere. That is pretty much how we use it currently, so maybe we'll just stick with this.

I think it also only feels wrong in German. Not sure. Will live with it a little and see if it bothers me.