Cleaned up / refactored SwitchMaster (before it even works)
I put all of the remaps analysis into separate methods. Now the "Update State" section is super readable and semantic and doesn't have all the low level remaps analysis going on.
Should eventually put these methods into RemapsAnalyzer

Also put device-info-getting methods into deviceManager, and applied some caching there so it costs close to nothing to retrieve that info.

Also improved some of the naming. There are now 3 abstract mouse 'features' that the SwitchMaster talks about: "Scroll", "Buttons" and "Pointing". And the naming of all the vars is consistent with that.

Reordered some stuff in DeviceManager and SwitchMaster to make it easier to navigate.

I should now finally start testing the code because none if this is going to work.