After removing the caching from `modificationsModifyButtons:`, the CPU usage is still 0.6 - 0.8%.
Caching seems to be not worth it here.

Edit:

Tested again against the old Swift code and this does seem a little bit faster. I guess the increases in CPU usage vs the older tests are because I click faster and faster? Or the computer has been running too long or sth? Idk.