Edit: I accidentally amended this commit instead o creating a new one:
Changed "Trackpad Interactions" to "Trackpad Simulation"

Now we finally have a solution that sounds good, is concise, and reasonably clear about the things it does why people might want to toggle it. (Scroll bouncing annoys some people, especially with a free-spinning scrollwheel and horizontal scroll between pages also annoys some people)

Also refactored NSAttributedString manipulation, added lots of docs and probably more I forgot.

----

Original commit

Changed "Advanced Navigation" to "Trackpad Interactions".

Not happy with it. The problem I'm really trying to solve is that I want people to understand it turns off scroll-bouncing, since that is the main reason why people might want to turn it off.  But I also want to mention that it lets you scroll between pages in Safari since that's a cool feature.

But I don't really know how to explain

 the scroll bouncing in a hint text. And also mentioning the bouncing and the scroll between pages in the hint text together would take up too much space.

So I guess if we rename to "trackpad interactions", that doesn't sound as good, but at least it makes it somewhat plausible to expect that this option turns off the scroll-bouncing.

I think it doesn't sound as good as "Advanced Navigation" and I don't like mentioning the trackpad directly in feature names. Because you should be thinking about what to do with it and how it works in and of itself. The fact that it works similar to a trackpad is more of an implementation detail.

But in this case I really don't know how to communicate all the things that this feature does in a concise manner except for "makes things work like a trackpad".

There is also no space in the hint for saying "makes things work like a trackpad" since we're already mentioning the safari page navigation.

And I really like giving an example of one of the useful things it actually does, instead of "JUST" saying "it makes things work like a trackpad", because that's very abstract and might be unclear.

So calling it "Trackpad Interactions" and then mentioning the scroll between pages in the hint is the best solution I can come up with for now. Even though it's not perfect.

Edit: I just reallllyyy dislike the name "Trackpad Interactions" vs "Advanced Navigation". I don't understand why, though. May go back if I dislike it too much.