Reworked GlobalEventTapThread into InputThread, and using the InputThread for processing button input and mouseMoved input.
Doing both on the same thread helps with some unreliability when triggering Click and Drag Gestures (especially when the computer is slow) I think. Before the events would sometimes be processed in the wrong order.

So far I don't see any bugs occuring from this. Except that twoFingerDrag doesn't work directly after threeFingerDrag. This might have been happening before already though. Not sure.

We should also analyze if this change makes a difference to performance and we should move the scroll input processing to the input thread as well.

For the processing of the inputs beyond the point where the different input types (click, drag, scroll) need to communicate / coodinate each other, it might be ideal to have a separate "outputThread" so the inputThread doesn't get overloaded. Maybe we could have a separate outputThread per output type, but that's probably overkill.