Revert "Reworked GlobalEventTapThread into InputThread, and using the InputThread for processing button input and mouseMoved input."
This reverts commit 2e4b7b5d8575c360cab7fc80471da2fd07b6df03.

There was a weird flicker when using twoFingerSwipe and releasing the button while moving the mouse. No idea why. But it's just too dangerous and too much work to change threading stuff right now.

I should make this change (Moving to a single input thread to make sure the order of event processing between button, scroll and drag input is correct) eventually. But not now.