In commit 575e2656cc824a659251ea2913ba39ba7605cca0 fixed a bug where dockSwipes driven by click and scroll would be canceled when the click and scroll button was released. If scroll-wheel event interception was **OFF** when the button was released.
But this introduced another bug where dockSwipes driven by click and scroll would be canceled when the click and scroll button was released. If scroll-wheel event interception was **ON** when the button was released. (Which should be the case for the configurations of most users.)

Now I think we fixed both of those bugs, and we also refactored and cleaned up the code around starting to receive and stopping to receive scrollwheel input.