Pretty significant refactor of HelperState.swift and dependent code.
The new idea for the for HelperState is that there is only one main interface for providing information to HelperState.swift and that is giving it the latestCGEvent.

All the public methods for getting info from HelperState don't have arguments anymore and are instead dynamically derived from the latestCGEvent. (There is one exception for this currently, sometimes, the information we want to provide is not a CGEvent, but that's relatively rare) I think this clearer way of thinking about the interface of HelperState.swift really simplifies things. Later we can also use this structure to optimize HelperState to a great extent making all the state be calculated lazily based on the latestCGEvent and also caching the relevant info from the CGEvent to avoid redundant calculations.