Moved all the .strings and .stringsdict files over to .xcstrings files.
Discussion:

.xcstrings files have the advantage of:

- Being able to be edited by users in a visual editor instead of plaintext
- Being able to easily sort the strings by different criteria in the visual editor
- ? Having translations for all the different languages in one file. So you have fewer files, which are easier to find and keep track of for translators and me.
- Automatic tracking of string state. The state can be "TRANSLATED", "NEW", "STALE", or "NEEDS REVIEW". Xcode manages this automatically. For example, when the string in the source language changes, then the translated strings are automatically marked with "Needs Review", which is amazing.
- Xcode also automatically tracks the translation progress and shows a "48% complete" badge next to a language, which is really nice.
- Compared to our State of Localization, it's nice that the stringState is shown right in line with the string, instead of having a separate document for that.

I feel like all the problems that our custom python scripts and the 'State of Localization' tried to solve are also solved by the .xcstring features - in a better way! (The problems that the custom system tried to solve were: Sorting the strings, finding syntax errors in the strings, marking unused strings (STALE), marking strings that have not yet been translated (NEW), marking strings that might be outdated. (NEEDS REVIEW).

So therefore I feel very tempted to adopt this system!

Based on this, I'm thinking about abolishing the 'State of Localization' code, since Xcode automatically tracks the 'State' of .xcstrings files. The only thing I'm wondering about is how do we solve these issues for the localizable strings on the MMF website and the MMF GitHub Readmes? To track the state of those, we might want to keep the State of Localization system or use a Translation Management System (TMS) like CrowdIn for everything.

If we stick with the custom system (over TMS), then it would be ideal if we could add "State" metadata right next to each of the kv-pairs for the website, and if we could have the website strings be editable in a GUI, instead of plaintext. Maybe we could achieve this by using xliff files to store the translations?

A TMS system might provide an even easier user experience and automation opportunities for my workflows. Also perhaps a TMS can manage the translations for the MMF website and for the MMF app in a unified way. They also provide features like Glossaries which may be useful to improve translation quality.

The downside is that TMS systems often cost something (many have free plans for FOSS software but MMF doesn't have a classic open source license and is monetized so I feel like it's unlikely we will qualify for the free plans). Aside from cost, further possible problems are that perhaps, the user experience and our workflows would actually be more complicated compared to simply letting users download and upload xcloc / xliff files which they can edit with some app and then send to me again.

But I should look into TMS systems. If we adopt a TMS system, we might not want to switch to using .xcstrings files since it might be unnecessary or even complicated integration?