- Restructured things a little to allow taking screenshots of the helper app. - Added the helpers statusMenuItem to the localizationScreenshots.  - Added localization screenshots for the AuthorizeAccessibilityView sheet.  - Also cleaned up the testRunner code a bit with `MARK:`, and other small stuff  - Maybe other improvements to testRunner code and stuff I forgot.
I'd say we're done with the localization screenshot implementation now! We screenshot almost the entire UI, and the UX should be quite nice for localizers.

There are still some small issues, like we don't take screenshots of some hard-to-mock states, like inability to connect to the internet or the licenseServer returning an error. Also some of the screenshots are taken a very small time before a toast completely disappears so there's a faint overlay of the last shown toast in the screenshot. But I think this slight lack of polish is really ok for something like this. Overall, I'm happy with the results now.

The only thing we should still do is replace the __underscore__ emphasis with **asterisk**. The underscores break the localization screenshots but they also break for Chinese and other scenarios so we should just replace them across the board.

Update:
TODO: Move `restore-default-buttons-popover` into toastAndSheetTests