Moved the high-level, interface functions inside License.swift and TrialCounter.swift to async/await.
Specifically we made the different License.checkLicense() variants and LicenseConfig.get() async. (Maybe other I forgot) and we added Task.detached() and `await` keywords in all the places that these were called.

We also made quite a big refactor to _checkLicense() - instead of first defining a wrapUp() closure and then calling that later, we have a `do {}` statement which performs the 'first step', and which we can break out of - into the 'second step' code which is what the 'wrapUp()' closure used to do before. So now The order of the code matches the order of what is happening. I hope that makes things nicer to read.

We also made some improvments in LicenseConfig.swift that goes beyong making thing get() async: Inside get(), we cleaned up the structure. We download the licenseConfig directly into memory instead of loading it into a file and then reading the file. We have also have new error generation for that. And when loading the licenseConfig from the internet fails for any reason, we actually log the error that caused the failure.
Also, we renamed a function, restructured some stuff for clarity, and added/improved some comments

The only other thing we changed I can think of where the behavior might have changed is those places where we used `Task.detached(priority: .background, ...` -> The background priority might make it execute it with a delay compared to before. Update: I tested this as bit and there's no noticable delay.
Sidenote: Don't forget the `Task.detached(priority: (triggeredByUser ? .userInitiated : .background),` line when searching for uses of `priority: .background`.