Reordered stuff inside License.swift. Gumroad.getLicenseInfo() has been renamed gumroad_getLicenseInfo() and moved into its singular caller _checkLicense() as a nested function. The Gumroad class has been removed, since there are no more Gumroad-specific functions now.
Background:
We're doing this in anticipation of integrating gumroad_getLicenseInfo into checkLicense() more deeply (instead of just as a nested function.)
However, I tried to do that in one step and came across the following problem: When you move code around and change it at the same time, all the side-by-side diff tools that I'd like to use to review these sensitive changes kinda fall apart. That's because when you move a chunk of code around it won't be tracked as the same chunk of code anymore by the diff algorithms. I tried many different diff tools. Araxis Merge kinda has a way around this problem, since it lets you scroll both sides of the diff independently, letting you manually adjust the scroll-position so you can visually compare the before and after of the section of code even if the diff algorithm doesn't understand that it's the same piece of code.
The problem is that Araxis Merge doesn't have syntax highlighting for Swift and costs like $100. I also tried a lot of other diff viewers and none of them could scroll the two sides independently.
So I think the best solution is to just change my habits:
For critical changes, I should never reorder sections of code and change them in the *same* commit! Instead:
	1. Step: There should be one commit where I just reorder things and don't change the code - so we're confident we didn't break things in that commit.
	2. Step: And then, in a second commit, we actually make the changes to the code. Then the diff algorithms understand the changes properly, and we can nicely use Xcodes "Side By Side Comparison" diff viewer to review the changes and be reasonably confident that we didn't break anything.

This commit represents the "1. Step"