Relatively scary refactor of _checkLicense().
We now declare a `result: MFLicenseState` at the very start and then override it based on different sources: serverResponse, cache, fallbackValues, and specialConditions (such as freeCountry). All these sources can give us a different `LicenseState` and the job of `_checkLicense` is primarily to gather the `LicenseState`s from the different sources, and to manage which sources to check in which order.
I think this new structure is much clearer.

We also made a slight refactor to the new askLicenseServers() function which we introduced in the last commit: It now returns an MFLicenseState. In the next commits we plan to also have a cache interface and a fallbackValue interface which return MFLicenseState instances to make the logic even clearer.

- We also slightly simplified parts of askLicenseServers(), preparing it to be used to talk to our AWS API as well as Gumroad.
- We improved some docs here or there.

- If we did everything right, the behaviour shouldn't have changed, except for 2 very small changes in _checkLicense:
	- Changing `SecureStorage.get(License.key) as! String?` to `SecureStorage.get(License.key) as? String`, potentially preventing a crash in some edge-cases.
	- Adding one extra assert to `Validate checkLicenseKeyValidity result` section

Overall this commit is a bit scary, even though behavior shouldn't have changed, because almost every line of _checkLicense() was changed since we do the same things in a pretty different way now.