Relatively large refactor:
Split up LicenseConfig.swift into 2 parts - storage and the retrieval:
- Storage is handled by new MFLicenseConfig, which is an MFDataClass defined in License.h and License.m.
- Retrieval of MFLicenseConfig instances is handled by LicenseConfig.swift which now only has static methods.

The code for creating/initializing the licenseConfig datastructure is totally different now, and lead to pretty large refactor in LicenseConfig.swift code which feels dangerous. I'd say 80-90% of the code in LicenseConfig.swift has been changed/replaced.

Also made a lot of small tangential changes while we were at it: (probably not a great idea to have this all in one commit)

- Reordered content inside License.h and moreso License.m. Hopefully easier to work with now. Also added comments and stuff there. Hopefully no changes in behavior. (Not that I remember, but it's hard to track the changes since we reordered everything - bad idea.)
- Other small stuff like improved an error message inside MFDataClass, comments and docs added about caching MFLicenseState (in License.swift), probably other comments and docs added / changed.

Meta:

This new structure is 'cleaner' in a way because we're reusing the features of MFDataClass. (We're now also making use of the validated dictionary init which we introduced in the last commit). Also, all of the 3 data containers that we obtain from License.swift and pass around the program (licenseConfig, licenseState and trialState) are all defined in License.h / License.m in a uniform way as MFDataClass subclasses. Also we could now further unify the code for retrieving licenseConfig/licenseState from server/cache/fallback.

However I still feel a bit bad about this change, because practically, this change brings no benefit. And also I didn't review/think it through as thoroughly as the last commits. So there's a decent chance that this could introduce new bugs. Also doing everything through MFDataClass didn't even result in much less code, the increased abstraction might be less clear, and putting everything into License.h / License.m arguably makes things harder to read. Also, all the changes we made in earlier commits unified and cleaned up the architecture of the Licensing code in ways that will help us implement offline validation and support for the AWS backend, and future licenseTypes more easily, and it fixed some bugs. And we thought through / reviewed the difficult changes very thorougly. This commit on the other hand, doesn't bring practical benefits, and has the potential of introducing new bugs.

Why did I do it anyways? I felt some sort of compulsion to 'clean things up' by unifying them in this way even though it's not practically important. I already knew that I shouldn't do this because it's a non-essential change that could bring new issues. But I thought "I just wanna try it out." Similar to the validated dict/json init we implemented yesterday. I felt like I justtttt wanna try it out a bit and see where it goes. But then I ended up making several small changes and improvements all over, and it would've become a chore to go through all the changes and only revert the ones that are dangerous. I also did a bunch of testing of different scenarios, and I did introduce some bugs, so this was definitely a 'dangerous' refactor, but now everything seems to work fine again.

So it feels kinda bad that I did this in the first place but it would feel even worse to try and go back now. So I guess we just roll with this.

My next thing that I feel compelled to do is try to 'clean things up' by handling the retrieval from server/cache/fallback the same way for the MFLicenseConfig as we do for the MFLicenseState. Then I wanna move all that retrieval code into License.swift I don't know if this a great idea.

the original thing that I wanted to do that lead me to this commit is have all the code inside License.swift obtain the MFLicenseConfig independently rather than getting the MFLicenseConfig passed in through an argument. I thought that this would let us completely avoid any internet requests in the 'offline validation' case - which might give people a better feeling about the privacy aspect. (And perhapss GitHub/Microsoft could track these requests in some way? But we're just downloading a 3KB JSON file, so idk if there's anything 'trackable' about this for GitHub / Microsoft.)