Implemented offline validation!
And some stuff to help the implementation of offline validation:

- Split up functions for retrieving the MFLicenseState, MFLicenseConfig, and MFTrialState. This allows for retrieving the MFLicenseConfig only if necessary, preventing unnecessary internet connections. (We already did a large part of this split last commit iirc, but we did even more restructuring, and removed checkLicenseAndTrial() entirely in favor of using the lowerLevel GetLicenseConfig, GetTrialState, GetLicenseState methods directly for more granular control.)
- Refactored all the places that use MFLicenseConfig so they only retrieve the MFLicenseConfig only if absolutely necessary.
- In GetLicenseState.swift, we implemented offline hash-based cache validation which lets us avoid internet connections in the normal case, while ensuring data integrity and making tampering relatively difficult.

Files:
	(These are changes in specific files that I found noteworthy while reviewing all changes, and that aren't already covered in the higher-level change-description above.)

	GetLicenseConfig.swift:
		- Fixed one tiny programmer error when passing nil to _licenseConfigDictCache.set{} where it would log that setting nil is undefined and does nothing but then it would still do something. Now it actually just returns and does nothing.
	default_config.plist
		- Added (empty) License.licenseStateCacheHash field to the default_config.plist to document the new field. This field is used by the offline validation code in GetLicenseState.swift.
	MacAddress.m/.h
		- ^ These are newly added files that let us retrieve the mac address of the hardware we're running on. This is used as a UID for the current device, which is used as an input for our offline validation hash. We wrote the new MFDefer macro to make this code nicer to write. (See MFDefer.h) The mac address is never stored or sent anywhere so this should not compromise user privacy in any way.
	MFDataClass
		- To aid debugging we implemented a custom [MFDataClassBase -description] method which recursively prints the contents of an MFDataClass instance. While preventing infinite loops if there are circular references.
			([NSDictionary -description]) just infinite-loops if there are circular references, so maybe this was overkill... But frigging ChatGPT told me I should worry about this. I need to learn to be more skeptical about its advice.
		- Also fixed a small programmer error where our code tried to determine whether a typeEncoding had an invalid length
	SharedUtility.h
		-> To prevent infinite recursions in the new [MFDataClass -description] method, we implemented a handy `threadlocal()` macro which lets us create a mutable object that persists between invocations of a function on the same thread. (I like writing macros now, should I be worried?)
	MFCoding.m/.h
		- These are new files, containing MFEncode() and MFDecode() which are a wrapper around NSKeyedArchiver/NSKeyedUnarchiver. They are not used, yet but I'll probably adopt them in the future. I think I finally solved the enigma that is the NSCoder API, and now we have an easy and flexible interface that correctly handles all exceptions and errors, and lets us control secureCoding and outputFormat. (Which is very hard and confusing for me with the underlying API.)
	MFDefer.h
		We implemented a Go/Swift style defer-statement for C/ObjC called `MFDefer`. It's really handy whenever you need to manually free() or CFRelease() variables.
	LicenseSheetController.swift
		Made the self.isProcessing boolean flag - which we use as a simple way to prevent race-conditions - atomic. Not sure that makes a difference.
---

Other smaller stuff such as:

- Renamed a bunch of arguments and variables for better clarity. Such as:
	- Calling it a license's 'activationCount' now instead of 'usesCount'
	- Renaming the licensing functions from ...Offline() to ..._Preliminary() to better reflect their purpose.
	- ...
- Added / improved lots of comments in the code.
	- Notably:
		In licenseConfigFromServer() we wrote a detailed breakdown of our architecture that enables offline validation.
		In MFCoding.m we wrote a detailed, 40 line explanation of how error handling works for NSCoder and why it ended up being so weird and confusing.
- Improved debug logs around the retrieval of MFLicenseState and MFLicenseConfig, aided by the new, improved `[MFDataClass -description]`
	I think this whole licensing code might be prone to rare, hard-to-reproduce bugs, so hopefully these logs will be helpful. Especially once we make it easier for users to acquire and send logs.

- Maybe other stuff I forgot.

-> Overall, the changes in this commit are pretty difficult, but I spent sooooo much time thinking about and reviewing everything, that I'm pretty confident it all works.

---

Meta:

Working on this commit has been soooo slow. The actual core offline validation implementation was pretty quick and fun to implement, but then I was like 'this is critical code, I now need to review every line thoroughly before I commit' and instead I sorta procrastinated by going down this rabbit hole of trying to put the deviceUID in the offline validation hash and then writing a defer macro in C to make obtaining the deviceUID nicer, and reading papers on the internet about adding a defer statement to the C standard, and then I wanted to make the cache human readable so it doesn't seem as ominous to users and to do that I needed to use the confusing parts of the NSCoder API, and then became obsessed with NSCoder and actually understanding how to use it, which took over a full day of work, and all this crazy stuff... None of that stuff was really necessary. I guess I thought it would be neat and interesting to do, and perhaps increase robustness and transparency for users a bit, but it just felt like such a draggg and waste of time in the end.

Also I'm being sooo thorough with these changes. It's nice in a way but it's not fun and I move so slow and it burns me out a bit, and then I get bored and go on tangents learning about stuff like using compiler attributes in macros to write a defer statement in C... And I wonder if this level of thoroughness and double-checking everything is really the best approach? I wonder if I could improve my attitude or workflow so that I could have high confidence that this critical code is correct without it becoming such a drag. Perhaps I should use unit testing after all? But I don't really think that would've helped here. Perhaps I should just throw out beta software more easily and improve the bug reporting tools so that I get great feedback from users who stumble across the edge-cases. Might be more productive in the long run. Not sure.. Will keep thinking about this.