Removed SnapKit mention from files in Markdown/Templates/*.
Next time we compile the docs, the reference should be removed in the output files, and the .xcstrings files should be updated.

Context:
We just removed SnapKit dependence on master branch in commit 25cdc1cbed61d4f3bcf6eb1525cace24b075c369.
The feature-strings-catalog branch (where we're committing this) still has the SnapKit dependence currently, but eventually we'll merge master.