Corrected event-sequence when gesture scrolls are canceled.
Before, when scrollwheel scrolling was canceled in the gesture phase (as opposed to the momentum phase), then we produced a `kIOHIDEventPhaseCancelled` event with (0,0)
deltas, but this event was then filtered out because the event-sending-code only expects zero deltas on kIOHIDEventPhaseEnded events.

This was incorrect and I believe caused janky behavior in some cases when canceling scrolls.
(I'm mostly thinking of iPad apps on Apple Silicon Macs, where cancelling sometimes didn't seem to work – but I can't test that on this machine.)

We've now fixed this such that, instead of a kIOHIDEventPhaseCancelled event, a kIOHIDEventPhaseEnded event is produced, which is followed
by a kCGMomentumScrollPhaseBegin and kCGMomentumScrollPhaseEnd event. This mirrors trackpad output relatively closely. It also mirrors old, deactivated code inside
cancel(forAutoMomentumScroll:), which we had already spent some time testing.

I also added/edited some comments and asserts() documenting our improved understanding.

(Context: Canceling is a nice feature that you can invoke by scrolling one tick opposite of the current scroll direction.)