This is an update to commit e588fb3426cdefd12710a13ef63a5023caf894f7
Instead of sending kCGMomentumScrollPhaseBegin and kCGMomentumScrollPhaseEnd events with zero-deltas to stop auto-momentumScroll, we're now sending kIOHIDEventPhaseMayBegin and kIOHIDEventPhaseCancelled events.
I think of this as simulating a tap on the trackpad.

This makes cancellation work reliably when running DownPay iPad app on Apple Silicon Mac. (Under the previous commit it didnt't work.)

I also added this tap-simulation to other places to make cancellation work correctly in all circumstances during scrollwheel-scrolling and Click-and-Drag-scrolling (hopefully)

We also updated/added comments to document and explain these decisions.