LocalizationScreenshotTaker.swift > testTakeScreenshots_Documentation() works really well now and produces good, annotated screenshots for all the languages we're interested in!
It's also pretty fast to iterate on now since it is capable of keeping the automated app running between runs of the testRunner.

We might tweak the details of the visual output it produces, but it's mostly done now.

The codepath of testTakeScreenshots_Localization() has also been affected by some refactors, although its behavior shouldn't have changed if we did everything correctly.