- Manually \escaped all the double-quotes in the xcstrings files that translate the .md files, so that they are valid JSON again. 	- Could have made a mistake. (Entered random extra characters) - Ran ./run syncstrings. (Not sure that did anything, but that what we used to determine which JSON files were corrupted.) - Didn't run ./run build-markdown, yet.
Discussion:

This mostly affects the Czech translator who used many smart-quotes in the md-file translations.
They used the smart-quotes even though they didn't use Xcode to edit the .xcloc files, which forces you to use smart-quotes. (They used Xcloc Editor.app, I'm pretty sure).
That suggests, that they made a deliberate choice to use smart-quotes, which I'm overriding here.
Maybe I should check back with them to ask if they're ok with that, and if not, write a script that adds smart-quotes back to their Czech translations.

I'm not sure all this is a good idea. I know that for English and German, the smart-quotes were only ever an accident, and they may make things harder to grep for. But for the translations, I'm not sure this is a benefit.
(It was just easier to do this way)